# apps/backend/src/project_backend/generated/http/models.py
# generated by datamodel-codegen:
#   filename:  openapi.yaml

from __future__ import annotations

from enum import Enum
from typing import Annotated

from pydantic import AnyUrl, AwareDatetime, BaseModel, ConfigDict, Field


class Status(Enum):
    ok = 'ok'


class Health(BaseModel):
    status: Annotated[Status, Field(examples=['ok'])]
    version: Annotated[str, Field(examples=['1.0.0'])]


class Status1(Enum):
    ready = 'ready'
    not_ready = 'not_ready'


class Readiness(BaseModel):
    status: Annotated[Status1, Field(examples=['ready'])]
    checks: Annotated[
        dict[str, str], Field(examples=[{'database': 'ok', 'redis': 'ok'}])
    ]


class Resource(BaseModel):
    id: Annotated[str, Field(examples=['res_01HXYZ'])]
    name: Annotated[
        str, Field(examples=['Summer sheet set'], max_length=120, min_length=1)
    ]
    locked: Annotated[bool, Field(examples=[False])]
    created_at: AwareDatetime


class ResourceCreate(BaseModel):
    model_config = ConfigDict(
        extra='forbid',
    )
    name: Annotated[str, Field(max_length=120, min_length=1)]


class ResourceList(BaseModel):
    items: list[Resource]
    total: Annotated[int, Field(ge=0)]
    page: Annotated[int, Field(ge=1)]
    page_size: Annotated[int, Field(ge=1)]


class Problem(BaseModel):
    type: AnyUrl
    title: str
    status: int
    code: Annotated[str, Field(description='Stable public error code')]
    detail: str | None = None
    instance: str | None = None
    trace_id: str | None = None